+++ Sources/QuickLMDB/LMDBError.swift
@@ -1,5 +1,4 @@
 import CLMDB
-import Foundation
 
 public enum LMDBError:Error {